protecting listing routes

+++ controllers/listingControllers.js
@@ -42,10 +42,19 @@ const getListingById = wrapAsync(async (req, res, next) => {
 });
 
 const createListing = (req, res) => {
-  res.render("listings/new");
+  if (req.isAuthenticated()) {
+    return res.render("listings/new");
+  }
+  req.flash("error", "You need to be logged in to create a listing");
+  res.redirect("/login");
 };
 
 const postListing = wrapAsync(async (req, res, next) => {
+  if (!req.isAuthenticated()) {
+    req.flash("error", "You need to be logged in to create a listing");
+    return res.redirect("/login");
+  }
+
   const { title, description, price, location, country } = req.body;
 
   validateListing(req.body);
@@ -69,6 +78,11 @@ const postListing = wrapAsync(async (req, res, next) => {
 });
 
 const editListing = wrapAsync(async (req, res, next) => {
+  if (!req.isAuthenticated()) {
+    req.flash("error", "You need to be logged in to edit a listing");
+    return res.redirect("/login");
+  }
+
   const { id } = req.params;
   const listing = await Listing.findById(id);
   if (!listing) {
@@ -80,6 +94,11 @@ const editListing = wrapAsync(async (req, res, next) => {
 });
 
 const updateListing = wrapAsync(async (req, res, next) => {
+  if (!req.isAuthenticated()) {
+    req.flash("error", "You need to be logged in to edit a listing");
+    return res.redirect("/login");
+  }
+
   const { id } = req.params;
   const { title, description, price, location, country, image } = req.body;
 
@@ -99,6 +118,11 @@ const updateListing = wrapAsync(async (req, res, next) => {
 });
 
 const deleteListing = wrapAsync(async (req, res, next) => {
+  if (!req.isAuthenticated()) {
+    req.flash("error", "You need to be logged in to delete a listing");
+    return res.redirect("/login");
+  }
+
   const { id } = req.params;
 
   const deletedListing = await Listing.findByIdAndDelete(id);
